refactor(test): move test utility functions to src/TestUtils.js
테스트 코드에서 많이 사용하는 mockQuestions, mockRandoms, getLogSpy 함수를 아예 다른 폴더로 구성해서 중복을 줄임.
