+++ __tests__/ApplicationTest.js
@@ -1,27 +1,5 @@
-const App = require("../src/App");
-const MissionUtils = require("@woowacourse/mission-utils");
-
-const mockQuestions = (answers) => {
-  MissionUtils.Console.readLine = jest.fn();
-  answers.reduce((acc, input) => {
-    return acc.mockImplementationOnce((question, callback) => {
-      callback(input);
-    });
-  }, MissionUtils.Console.readLine);
-};
-
-const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickNumberInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickNumberInRange);
-};
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, "print");
-  logSpy.mockClear();
-  return logSpy;
-};
+const App = require('../src/App');
+const { mockQuestions, mockRandoms, getLogSpy } = require('../src/TestUtils');
 
 describe("숫자 야구 게임", () => {
   test("게임 종료 후 재시작 1", () => {
