+++ __tests__/SingleGameTest.js
@@ -1,27 +1,5 @@
-const MissionUtils = require('@woowacourse/mission-utils');
 const App = require('../src/App');
-
-const mockQuestions = (answers) => {
-  MissionUtils.Console.readLine = jest.fn();
-  answers.reduce((acc, input) => {
-    return acc.mockImplementationOnce((question, callback) => {
-      callback(input);
-    });
-  }, MissionUtils.Console.readLine);
-};
-
-const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickNumberInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickNumberInRange);
-};
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
-  logSpy.mockClear();
-  return logSpy;
-};
+const { mockQuestions, mockRandoms, getLogSpy } = require('../src/TestUtils');
 
 describe('숫자 야구 게임 단판', () => {
   test('단판전 1', () => {
