+++ src/TestUtils.js
@@ -1,0 +1,25 @@
+const MissionUtils = require('@woowacourse/mission-utils');
+
+const mockQuestions = (answers) => {
+  MissionUtils.Console.readLine = jest.fn();
+  answers.reduce(
+    (acc, input) => acc.mockImplementationOnce((question, callback) => { callback(input); }),
+    MissionUtils.Console.readLine,
+  );
+};
+
+const mockRandoms = (numbers) => {
+  MissionUtils.Random.pickNumberInRange = jest.fn();
+  numbers.reduce(
+    (acc, number) => acc.mockReturnValueOnce(number),
+    MissionUtils.Random.pickNumberInRange,
+  );
+};
+
+const getLogSpy = () => {
+  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
+  logSpy.mockClear();
+  return logSpy;
+};
+
+module.exports = { mockQuestions, mockRandoms, getLogSpy };
